% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_extrapolation.R
\name{meps_extrapolate_tail}
\alias{meps_extrapolate_tail}
\title{Impute Ranks for Truncated Upper Tails Using an Age-Conditional Reference}
\usage{
meps_extrapolate_tail(
  meps_data,
  reported_var,
  truncation_val,
  qr_grid,
  tau_seq = c(seq(0.01, 0.95, by = 0.01), seq(0.96, 0.999, by = 0.001))
)
}
\arguments{
\item{meps_data}{Cleaned target MEPS dataset containing self-reported values.}

\item{reported_var}{Character string of the column name containing self-reported values.}

\item{truncation_val}{The numeric threshold where the survey implements top-coding.}

\item{qr_grid}{A fitted quantile grid object generated by `fit_age_spline_quantiles` on NHANES data.}

\item{tau_seq}{Numeric vector of quantiles matching the grid setup.}
}
\value{
A numeric vector of length nrow(meps_data) containing continuous, de-truncated U(0,1) percentiles.
}
\description{
Impute Ranks for Truncated Upper Tails Using an Age-Conditional Reference
}
