% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_meps.R
\name{meps_quantreg_calibrate}
\alias{meps_quantreg_calibrate}
\title{Calibrate Self-Reported MEPS Values into Continuous Age-Adjusted Copula Space}
\usage{
meps_quantreg_calibrate(
  meps_data,
  reported_var,
  qr_grid,
  tau_seq = c(seq(0.01, 0.95, by = 0.01), seq(0.96, 0.999, by = 0.001))
)
}
\arguments{
\item{meps_data}{Cleaned target MEPS dataset containing self-reported values.}

\item{reported_var}{Character string of the column name containing self-reported values.}

\item{qr_grid}{A fitted quantile grid object generated by `fit_age_spline_quantiles`.}

\item{tau_seq}{Numeric vector of quantiles matching the grid setup.}
}
\value{
A continuous numeric vector of length nrow(meps_data) containing de-heaped, age-conditional U(0,1) percentiles.
}
\description{
Calibrate Self-Reported MEPS Values into Continuous Age-Adjusted Copula Space
}
